restructure testing directories and classed
