+++ phpunit.xml
@@ -22,12 +22,6 @@
         <testsuite name="LaraSearch Test Suite">
             <directory>tests</directory>
         </testsuite>
-        <testsuite name="Unit">
-            <directory suffix="Test.php">./tests/Unit</directory>
-        </testsuite>
-        <testsuite name="Feature">
-            <directory suffix="Test.php">./tests/Feature</directory>
-        </testsuite>
     </testsuites>
     <php>
         <env name="CACHE_DRIVER" value="array"/>
