--- tests/Unit/ExampleTest.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Muhaimenul\Larasearch\Tests\Unit;
-
-use PHPUnit\Framework\TestCase;
-
-class ExampleTest extends TestCase
-{
-    /**
-     * A basic test example.
-     *
-     * @return void
-     */
-    public function test_example()
-    {
-        $this->assertTrue(true);
-    }
-}
